[feat] Restaurant 관련 Repository, Service 클래스 생성
