+++ src/main/java/team/waitingcatch/app/common/Address.java
@@ -1,0 +1,19 @@
+package team.waitingcatch.app.common;
+
+import lombok.Getter;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+
+@Embeddable
+@Getter
+public class Address {
+    @Column(nullable = false)
+    private String province;
+
+    @Column(nullable = false)
+    private String city;
+
+    @Column(nullable = false)
+    private String street;
+}
