--- src/main/java/team/waitingcatch/app/restaurant/entity/Address.java
@@ -1,12 +1,0 @@
-package team.waitingcatch.app.restaurant.entity;
-
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class Address {
-    private String province;
-    private String city;
-    private String street;
-    private float latitude;
-    private float longitude;
-}
