+++ src/main/java/team/waitingcatch/app/restaurant/repository/MenuRepository.java
@@ -1,0 +1,7 @@
+package team.waitingcatch.app.restaurant.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import team.waitingcatch.app.restaurant.entity.Menu;
+
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+}
