+++ src/main/java/team/waitingcatch/app/restaurant/repository/RestaurantCategoryRepository.java
@@ -1,0 +1,7 @@
+package team.waitingcatch.app.restaurant.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import team.waitingcatch.app.restaurant.entity.RestaurantCategory;
+
+public interface RestaurantCategoryRepository extends JpaRepository<RestaurantCategory, Long> {
+}
