+++ src/main/java/team/waitingcatch/app/restaurant/service/category/CategoryService.java
@@ -1,0 +1,4 @@
+package team.waitingcatch.app.restaurant.service.category;
+
+public interface CategoryService {
+}
