+++ src/main/java/team/waitingcatch/app/restaurant/service/category/CategoryServiceImpl.java
@@ -1,0 +1,4 @@
+package team.waitingcatch.app.restaurant.service.category;
+
+public class CategoryServiceImpl implements CategoryService, InternalCategoryService{
+}
