+++ src/main/java/team/waitingcatch/app/restaurant/service/menu/MenuService.java
@@ -1,0 +1,4 @@
+package team.waitingcatch.app.restaurant.service.menu;
+
+public interface MenuService {
+}
