+++ src/main/java/team/waitingcatch/app/restaurant/service/menu/MenuServiceImpl.java
@@ -1,0 +1,4 @@
+package team.waitingcatch.app.restaurant.service.menu;
+
+public class MenuServiceImpl implements MenuService, InternalMenuService{
+}
