+++ src/main/java/team/waitingcatch/app/restaurant/service/restaurant/InternalRestaurantService.java
@@ -1,0 +1,4 @@
+package team.waitingcatch.app.restaurant.service.restaurant;
+
+public interface InternalRestaurantService {
+}
