+++ src/main/java/team/waitingcatch/app/restaurant/service/restaurant/RestaurantServiceImpl.java
@@ -1,0 +1,4 @@
+package team.waitingcatch.app.restaurant.service.restaurant;
+
+public class RestaurantServiceImpl implements RestaurantService, InternalRestaurantService{
+}
